Update api.py
Updated the function for uploading instructions to allow for CSV's, JSON or JSONL. Also allowed for a local file or a Hugging Face dataset. If the HF dataset is private I allowed for the optional param of hf_token.

I also force the dataset to have one of the following structure:
required_columns_sets = [
        {'prompt', 'messages'},
        {'question', 'answer'},
        {'prompt', 'completion'},
        {'instruction', 'response'}
    ]
as well as a optional 3rd column for 'split' where the dataset can have a train or test value. This will allow for a split based on these values before training over us doing a random split.